Replace most chars and ints with uint*_t

+++ src/encparams.c
@@ -1,17 +1,17 @@
 #include "encparams.h"
 
-int ntru_enc_len(int N, int q) {
+uint16_t ntru_enc_len(uint16_t N, uint16_t q) {
     /* make sure q is a power of 2 */
     if (q & (q-1))
         return -1;
 
-    int log2q = 0;
+    uint16_t log2q = 0;
     while (q > 1) {
         q /= 2;
         log2q++;
     }
 
-    int len_bits = N * log2q;
-    int len_bytes = (len_bits+7) / 8;
+    uint16_t len_bits = N * log2q;
+    uint16_t len_bytes = (len_bits+7) / 8;
     return len_bytes;
 }
