+++ tests/test_hash.h
@@ -1,6 +1,8 @@
 #ifndef TEST_HASH_H
 #define TEST_HASH_H
 
-int test_hash();
+#include <stdint.h>
+
+uint8_t test_hash();
 
 #endif
